MYR-1802: changed endpoint to check for nonce
