+++ src/lib/api/user.ts
@@ -13,19 +13,6 @@ import {
 type UserList = BaseList<User>;
 type ActivityList = BaseList<ActivityLog>;
 
-type UserNonceProps = {
-  nonce: number;
-};
-
-export const getUserNonce = async (id: string): Promise<UserNonceProps> => {
-  const {data} = await MyriadAPI.request({
-    url: `users/${id}/nonce`,
-    method: 'GET',
-  });
-
-  return data ? data : {nonce: 0};
-};
-
 export const getUserDetail = async (id: string, userId?: string): Promise<User & BlockedProps> => {
   const params: Record<string, any> = {
     filter: {
